Update sequelize config
