--- api/config/db.js
@@ -1,23 +1,0 @@
-require('dotenv').config();
-const { Sequelize } = require("sequelize");
-
-const sequelize = new Sequelize(
-  process.env.DB_NAME, 
-  process.env.DB_USER, 
-  process.env.DB_PASSWORD, 
-  {
-    dialect: 'postgres',
-    host: process.env.DB_HOST
-  }
-)
-
-const testDatabaseConnection = async () => {
-  try {
-    await sequelize.authenticate();
-    console.log("Connection has been established successfully.");
-  } catch (error) {
-    console.error("Unable to connect to the database:", error);
-  }
-};
-
-module.exports = { sequelize, testDatabaseConnection };
